refactor(dropdown): take out dropdown from the currency selector
close #46
