+++ src/components/CurrencySelector/CurrencySelector.scss
