+++ src/components/Dropdown/Dropdown.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './Dropdown.scss';
+
+type DropdownProps = { children: React.ReactNode };
+
+const Dropdown = ({ children }: DropdownProps): JSX.Element => {
+  return <div className="dropdown">{children}</div>;
+};
+
+export default Dropdown;
