+++ src/components/Dropdown/DropdownList/index.ts
@@ -1,0 +1,3 @@
+import DropdownList from './DropdownList';
+
+export default DropdownList;
